LegalMitra/requirements.txt
streamlit==1.32.0
PyMuPDF==1.23.8
google-generativeai==0.3.2
python-dotenv==1.0.0